requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.0
decorator==4.4.2
dill==0.3.2
etelemetry==0.2.1
filelock==3.0.12
fpdf==1.7.2
httplib2==0.18.1
idna==2.10
img2pdf==0.3.6
iniconfig==1.0.1
isodate==0.6.0
lxml==4.5.2
more-itertools==8.4.0
networkx==2.4
nibabel==3.1.1
nipype==1.5.0
numpy==1.19.1
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
pkg-resources==0.0.0
pluggy==0.13.1
prov==1.5.3
py==1.9.0
pydot==1.4.1
pydotplus==2.0.2
pyfunctional==1.4.1
PyMuPDF==1.17.4
pyparsing==2.4.7
PyPDF4==1.27.0
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
pyxnat==1.3
pyzbar==0.1.8
qrcode==6.1
rdflib==5.0.0
requests==2.24.0
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
tabulate==0.8.7
toml==0.10.1
tqdm==4.48.0
traits==6.1.1
urllib3==1.25.10
